Issue #2807: Move ExperimentPage container into a bundle directory, move IncompatibleAddons subcomponent into the bundle
