--- frontend/src/app/components/IncompatibleAddons/index.scss
@@ -1,2 +1,0 @@
-// TODO: migrate <IncompatibleAddons> styles from frontend/src/styles/components/_Warning.scss
-
